+++ DSCResources/MSFT_FirewallProfile/MSFT_FirewallProfile.psm1
@@ -5,15 +5,8 @@ Import-Module -Name (Join-Path -Path $modulePath `
                                -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
                                                      -ChildPath 'NetworkingDsc.Common.psm1'))
 
-# Import the Networking Resource Helper Module
-Import-Module -Name (Join-Path -Path $modulePath `
-                               -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
-                                                     -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
-
 # Import Localization Strings
-$LocalizedData = Get-LocalizedData `
-    -ResourceName 'MSFT_FirewallProfile' `
-    -ResourcePath (Split-Path -Parent $Script:MyInvocation.MyCommand.Path)
+$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_FirewallProfile'
 
 <#
     This is an array of all the parameters used by this resource.
@@ -46,7 +39,7 @@ function Get-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($LocalizedData.GettingFirewallProfileMessage) `
+            $($script:localizedData.GettingFirewallProfileMessage) `
                 -f $Name
         ) -join '' )
 
@@ -231,7 +224,7 @@ function Set-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($LocalizedData.SettingFirewallProfileMessage) `
+            $($script:localizedData.SettingFirewallProfileMessage) `
                 -f $Name
         ) -join '' )
 
@@ -256,7 +249,7 @@ function Set-TargetResource
 
             Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($LocalizedData.FirewallProfileUpdateParameterMessage) `
+                $($script:localizedData.FirewallProfileUpdateParameterMessage) `
                     -f $Name,$parameter.Name,$parameterNewValue
                 ) -join '' )
         } # if
@@ -271,7 +264,7 @@ function Set-TargetResource
 
         Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($LocalizedData.FirewallProfileUpdatedMessage) `
+            $($script:localizedData.FirewallProfileUpdatedMessage) `
                 -f $Name
             ) -join '' )
     } # if
@@ -439,7 +432,7 @@ function Test-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($LocalizedData.TestingFirewallProfileMessage) `
+            $($script:localizedData.TestingFirewallProfileMessage) `
                 -f $Name
         ) -join '' )
 
@@ -461,7 +454,7 @@ function Test-TargetResource
         {
             Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($LocalizedData.FirewallProfileParameterNeedsUpdateMessage) `
+                $($script:localizedData.FirewallProfileParameterNeedsUpdateMessage) `
                     -f $Name,$parameter.Name,$parameterSourceValue,$parameterNewValue
                 ) -join '' )