+++ DSCResources/MSFT_IPAddressOption/MSFT_IPAddressOption.psm1
@@ -5,17 +5,8 @@ Import-Module -Name (Join-Path -Path $modulePath `
         -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
             -ChildPath 'NetworkingDsc.Common.psm1'))
 
-# Import the Networking Resource Helper Module
-Import-Module -Name (Join-Path -Path $modulePath `
-        -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
-            -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
-
 # Import Localization Strings
-$localizedDataSplat = @{
-    ResourceName = 'MSFT_IPAddressOption'
-    ResourcePath = (Split-Path -Parent $Script:MyInvocation.MyCommand.Path)
-}
-$localizedData = Get-LocalizedData @localizedDataSplat
+$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_IPAddressOption'
 
 <#
     .SYNOPSIS
@@ -44,7 +35,7 @@ function Get-TargetResource
     )
 
     Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
-            $($LocalizedData.GettingIPAddressOptionMessage -f $IPAddress)
+            $($script:localizedData.GettingIPAddressOptionMessage -f $IPAddress)
         ) -join '')
 
     $currentIPAddress = Get-NetIPAddress -IPAddress $IPAddress
@@ -83,7 +74,7 @@ function Set-TargetResource
     )
 
     Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
-            $($LocalizedData.ApplyingIPAddressOptionMessage -f $IPAddress)
+            $($script:localizedData.ApplyingIPAddressOptionMessage -f $IPAddress)
         ) -join '')
 
     $currentConfig = Get-TargetResource @PSBoundParameters
@@ -124,7 +115,7 @@ function Test-TargetResource
     [System.Boolean] $desiredConfigurationMatch = $true
 
     Write-Verbose -Message ( @("$($MyInvocation.MyCommand): "
-            $($LocalizedData.CheckingIPAddressOptionMessage -f $IPAddress)
+            $($script:localizedData.CheckingIPAddressOptionMessage -f $IPAddress)
         ) -join '')
 
     $currentConfig = Get-TargetResource @PSBoundParameters