+++ DSCResources/MSFT_NetAdapterAdvancedProperty/MSFT_NetAdapterAdvancedProperty.psm1
@@ -5,15 +5,8 @@ Import-Module -Name (Join-Path -Path $modulePath `
         -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
             -ChildPath 'NetworkingDsc.Common.psm1'))
 
-# Import the Networking Resource Helper Module
-Import-Module -Name (Join-Path -Path $modulePath `
-        -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
-            -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
-
 # Import Localization Strings
-$localizedData = Get-LocalizedData `
-    -ResourceName 'MSFT_NetAdapterAdvancedProperty' `
-    -ResourcePath (Split-Path -Parent $Script:MyInvocation.MyCommand.Path)
+$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_NetAdapterAdvancedProperty'
 
 <#
 .SYNOPSIS
@@ -39,7 +32,7 @@ function Get-TargetResource
         $NetworkAdapterName,
 
         [Parameter(Mandatory = $true)]
-        [ValidateSet("*DcbxMode","*EncapsulatedPacketTaskOffload","*FlowControl","*InterruptModeration","*IPChecksumOffloadIPv4","*JumboPacket","*LsoV2IPv4","*LsoV2IPv6","*MaxRssProcessors","*NetworkDirect","*NumaNodeId","*NumRssQueues","*PacketDirect","*PriorityVLANTag","*QOS","*ReceiveBuffers","*RecvCompletionMethod","*RoceMaxFrameSize","*RscIPv4","*RSS","*RssBaseProcNumber","*RssMaxProcNumber","*RssOnHostVPorts","*RSSProfile","RxIntModeration","RxIntModerationProfile","*SpeedDuplex","*Sriov","*TCPChecksumOffloadIPv4","*TCPChecksumOffloadIPv6","*TCPUDPChecksumOffloadIPv4","*TCPUDPChecksumOffloadIPv6","*TransmitBuffers","TxIntModerationProfile","*UDPChecksumOffloadIPv4","*UDPChecksumOffloadIPv6","VlanID","*VMQ","*VMQVlanFiltering","AdaptiveIFS","ITR","LogLinkStateEvent","MasterSlave","NetworkAddress","WaitAutoNegComplete")]
+        [ValidateSet('*DcbxMode','*EncapsulatedPacketTaskOffload','*FlowControl','*InterruptModeration','*IPChecksumOffloadIPv4','*JumboPacket','*LsoV2IPv4','*LsoV2IPv6','*MaxRssProcessors','*NetworkDirect','*NumaNodeId','*NumRssQueues','*PacketDirect','*PriorityVLANTag','*QOS','*ReceiveBuffers','*RecvCompletionMethod','*RoceMaxFrameSize','*RscIPv4','*RSS','*RssBaseProcNumber','*RssMaxProcNumber','*RssOnHostVPorts','*RSSProfile','RxIntModeration','RxIntModerationProfile','*SpeedDuplex','*Sriov','*TCPChecksumOffloadIPv4','*TCPChecksumOffloadIPv6','*TCPUDPChecksumOffloadIPv4','*TCPUDPChecksumOffloadIPv6','*TransmitBuffers','TxIntModerationProfile','*UDPChecksumOffloadIPv4','*UDPChecksumOffloadIPv6','VlanID','*VMQ','*VMQVlanFiltering','AdaptiveIFS','ITR','LogLinkStateEvent','MasterSlave','NetworkAddress','WaitAutoNegComplete','MaxRxRing1Length','NumRxBuffersSmall')]
         [System.String]
         $RegistryKeyword,
 
@@ -50,7 +43,7 @@ function Get-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $localizedData.CheckingNetAdapterMessage
+            $script:localizedData.CheckingNetAdapterMessage
         ) -join '')
 
     try
@@ -63,14 +56,14 @@ function Get-TargetResource
     catch
     {
         New-InvalidOperationException `
-            -Message ($LocalizedData.NetAdapterNotFoundMessage)
+            -Message ($script:localizedData.NetAdapterNotFoundMessage)
     }
 
     if ($netAdapterAdvancedProperty)
     {
         Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($LocalizedData.NetAdapterTestingStateMessage -f $NetworkAdapterName, $RegistryKeyword)
+                $($script:localizedData.NetAdapterTestingStateMessage -f $NetworkAdapterName, $RegistryKeyword)
             ) -join '')
 
         $result = @{
@@ -107,7 +100,7 @@ function Set-TargetResource
         $NetworkAdapterName,
 
         [Parameter(Mandatory = $true)]
-        [ValidateSet("*DcbxMode","*EncapsulatedPacketTaskOffload","*FlowControl","*InterruptModeration","*IPChecksumOffloadIPv4","*JumboPacket","*LsoV2IPv4","*LsoV2IPv6","*MaxRssProcessors","*NetworkDirect","*NumaNodeId","*NumRssQueues","*PacketDirect","*PriorityVLANTag","*QOS","*ReceiveBuffers","*RecvCompletionMethod","*RoceMaxFrameSize","*RscIPv4","*RSS","*RssBaseProcNumber","*RssMaxProcNumber","*RssOnHostVPorts","*RSSProfile","RxIntModeration","RxIntModerationProfile","*SpeedDuplex","*Sriov","*TCPChecksumOffloadIPv4","*TCPChecksumOffloadIPv6","*TCPUDPChecksumOffloadIPv4","*TCPUDPChecksumOffloadIPv6","*TransmitBuffers","TxIntModerationProfile","*UDPChecksumOffloadIPv4","*UDPChecksumOffloadIPv6","VlanID","*VMQ","*VMQVlanFiltering","AdaptiveIFS","ITR","LogLinkStateEvent","MasterSlave","NetworkAddress","WaitAutoNegComplete")]
+        [ValidateSet('*DcbxMode','*EncapsulatedPacketTaskOffload','*FlowControl','*InterruptModeration','*IPChecksumOffloadIPv4','*JumboPacket','*LsoV2IPv4','*LsoV2IPv6','*MaxRssProcessors','*NetworkDirect','*NumaNodeId','*NumRssQueues','*PacketDirect','*PriorityVLANTag','*QOS','*ReceiveBuffers','*RecvCompletionMethod','*RoceMaxFrameSize','*RscIPv4','*RSS','*RssBaseProcNumber','*RssMaxProcNumber','*RssOnHostVPorts','*RSSProfile','RxIntModeration','RxIntModerationProfile','*SpeedDuplex','*Sriov','*TCPChecksumOffloadIPv4','*TCPChecksumOffloadIPv6','*TCPUDPChecksumOffloadIPv4','*TCPUDPChecksumOffloadIPv6','*TransmitBuffers','TxIntModerationProfile','*UDPChecksumOffloadIPv4','*UDPChecksumOffloadIPv6','VlanID','*VMQ','*VMQVlanFiltering','AdaptiveIFS','ITR','LogLinkStateEvent','MasterSlave','NetworkAddress','WaitAutoNegComplete','MaxRxRing1Length','NumRxBuffersSmall')]
         [System.String]
         $RegistryKeyword,
 
@@ -118,7 +111,7 @@ function Set-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $localizedData.CheckingNetAdapterMessage
+            $script:localizedData.CheckingNetAdapterMessage
         ) -join '')
 
     try
@@ -131,14 +124,14 @@ function Set-TargetResource
     catch
     {
         New-InvalidOperationException `
-            -Message ($LocalizedData.NetAdapterNotFoundMessage)
+            -Message ($script:localizedData.NetAdapterNotFoundMessage)
     }
 
     if ($netAdapterAdvancedProperty)
     {
         Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($LocalizedData.NetAdapterTestingStateMessage -f $NetworkAdapterName, $RegistryKeyword)
+                $($script:localizedData.NetAdapterTestingStateMessage -f $NetworkAdapterName, $RegistryKeyword)
             ) -join '')
 
         if ($RegistryValue -ne $netAdapterAdvancedProperty.RegistryValue)
@@ -146,7 +139,7 @@ function Set-TargetResource
             $netadapterRegistryValue = $netAdapterAdvancedProperty.RegistryValue
             Write-Verbose -Message ( @(
                     "$($MyInvocation.MyCommand): "
-                    $($LocalizedData.NetAdapterApplyingChangesMessage -f `
+                    $($script:localizedData.NetAdapterApplyingChangesMessage -f `
                             $NetworkAdapterName, $RegistryKeyword, "$netadapterRegistryValue", $RegistryValue )
                 ) -join '')
 
@@ -182,7 +175,7 @@ function Test-TargetResource
         $NetworkAdapterName,
 
         [Parameter(Mandatory = $true)]
-        [ValidateSet("*DcbxMode","*EncapsulatedPacketTaskOffload","*FlowControl","*InterruptModeration","*IPChecksumOffloadIPv4","*JumboPacket","*LsoV2IPv4","*LsoV2IPv6","*MaxRssProcessors","*NetworkDirect","*NumaNodeId","*NumRssQueues","*PacketDirect","*PriorityVLANTag","*QOS","*ReceiveBuffers","*RecvCompletionMethod","*RoceMaxFrameSize","*RscIPv4","*RSS","*RssBaseProcNumber","*RssMaxProcNumber","*RssOnHostVPorts","*RSSProfile","RxIntModeration","RxIntModerationProfile","*SpeedDuplex","*Sriov","*TCPChecksumOffloadIPv4","*TCPChecksumOffloadIPv6","*TCPUDPChecksumOffloadIPv4","*TCPUDPChecksumOffloadIPv6","*TransmitBuffers","TxIntModerationProfile","*UDPChecksumOffloadIPv4","*UDPChecksumOffloadIPv6","VlanID","*VMQ","*VMQVlanFiltering","AdaptiveIFS","ITR","LogLinkStateEvent","MasterSlave","NetworkAddress","WaitAutoNegComplete")]
+        [ValidateSet('*DcbxMode','*EncapsulatedPacketTaskOffload','*FlowControl','*InterruptModeration','*IPChecksumOffloadIPv4','*JumboPacket','*LsoV2IPv4','*LsoV2IPv6','*MaxRssProcessors','*NetworkDirect','*NumaNodeId','*NumRssQueues','*PacketDirect','*PriorityVLANTag','*QOS','*ReceiveBuffers','*RecvCompletionMethod','*RoceMaxFrameSize','*RscIPv4','*RSS','*RssBaseProcNumber','*RssMaxProcNumber','*RssOnHostVPorts','*RSSProfile','RxIntModeration','RxIntModerationProfile','*SpeedDuplex','*Sriov','*TCPChecksumOffloadIPv4','*TCPChecksumOffloadIPv6','*TCPUDPChecksumOffloadIPv4','*TCPUDPChecksumOffloadIPv6','*TransmitBuffers','TxIntModerationProfile','*UDPChecksumOffloadIPv4','*UDPChecksumOffloadIPv6','VlanID','*VMQ','*VMQVlanFiltering','AdaptiveIFS','ITR','LogLinkStateEvent','MasterSlave','NetworkAddress','WaitAutoNegComplete','MaxRxRing1Length','NumRxBuffersSmall')]
         [System.String]
         $RegistryKeyword,
 
@@ -193,7 +186,7 @@ function Test-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $localizedData.CheckingNetAdapterMessage
+            $script:localizedData.CheckingNetAdapterMessage
         ) -join '')
 
     try
@@ -206,14 +199,14 @@ function Test-TargetResource
     catch
     {
         New-InvalidOperationException `
-            -Message ($LocalizedData.NetAdapterNotFoundMessage)
+            -Message ($script:localizedData.NetAdapterNotFoundMessage)
     }
 
     if ($netAdapterAdvancedProperty)
     {
         Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $localizedData.NetAdapterTestingStateMessage -f `
+                $script:localizedData.NetAdapterTestingStateMessage -f `
                     $NetworkAdapterName, $RegistryKeyword
             ) -join '')