+++ DSCResources/MSFT_NetAdapterName/MSFT_NetAdapterName.psm1
@@ -5,15 +5,8 @@ Import-Module -Name (Join-Path -Path $modulePath `
         -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
             -ChildPath 'NetworkingDsc.Common.psm1'))
 
-# Import the Networking Resource Helper Module
-Import-Module -Name (Join-Path -Path $modulePath `
-        -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
-            -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
-
 # Import Localization Strings
-$localizedData = Get-LocalizedData `
-    -ResourceName 'MSFT_NetAdapterName' `
-    -ResourcePath (Split-Path -Parent $Script:MyInvocation.MyCommand.Path)
+$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_NetAdapterName'
 
 <#
     .SYNOPSIS
@@ -111,7 +104,7 @@ function Get-TargetResource
     )
 
     Write-Verbose -Message ( @("$($MyInvocation.MyCommand): "
-            $($LocalizedData.GettingNetAdapterNameMessage -f $NewName)
+            $($script:localizedData.GettingNetAdapterNameMessage -f $NewName)
         ) -join '')
 
     $adapter = Find-NetworkAdapter `
@@ -121,7 +114,7 @@ function Get-TargetResource
     if (-not $adapter)
     {
         Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
-                $($LocalizedData.FindNetAdapterMessage)
+                $($script:localizedData.FindNetAdapterMessage)
             ) -join '')
 
         $null = $PSBoundParameters.Remove('NewName')
@@ -132,7 +125,7 @@ function Get-TargetResource
     }
 
     Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
-            $($LocalizedData.NetAdapterNameFoundMessage -f $adapter.Name)
+            $($script:localizedData.NetAdapterNameFoundMessage -f $adapter.Name)
         ) -join '')
 
     $returnValue = @{
@@ -245,7 +238,7 @@ function Set-TargetResource
     )
 
     Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
-            $($LocalizedData.SettingNetAdapterNameMessage -f $NewName)
+            $($script:localizedData.SettingNetAdapterNameMessage -f $NewName)
         ) -join '')
 
     $null = $PSBoundParameters.Remove('NewName')
@@ -255,13 +248,13 @@ function Set-TargetResource
         -ErrorAction Stop
 
     Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
-            $($LocalizedData.RenamingNetAdapterNameMessage -f $adapter.Name, $NewName)
+            $($script:localizedData.RenamingNetAdapterNameMessage -f $adapter.Name, $NewName)
         ) -join '')
 
     $adapter | Rename-NetAdapter -NewName $NewName
 
     Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
-            $($LocalizedData.NetAdapterNameRenamedMessage -f $NewName)
+            $($script:localizedData.NetAdapterNameRenamedMessage -f $NewName)
         ) -join '')
 } # Set-TargetResource
 
@@ -361,7 +354,7 @@ function Test-TargetResource
     )
 
     Write-Verbose -Message ( @("$($MyInvocation.MyCommand): "
-            $($LocalizedData.TestingNetAdapterNameMessage -f $NewName)
+            $($script:localizedData.TestingNetAdapterNameMessage -f $NewName)
         ) -join '')
 
     $null = $PSBoundParameters.Remove('NewName')
@@ -376,7 +369,7 @@ function Test-TargetResource
     {
         # An adapter was found matching the new name
         Write-Verbose -Message ( @("$($MyInvocation.MyCommand): "
-                $($LocalizedData.NetAdapterWithNewNameExistsMessage -f $adapterWithNewName.Name)
+                $($script:localizedData.NetAdapterWithNewNameExistsMessage -f $adapterWithNewName.Name)
             ) -join '')
 
         return $true
@@ -385,7 +378,7 @@ function Test-TargetResource
     {
         # Find an adapter matching the parameters - throw if none can be found
         Write-Verbose -Message ( @( "$($MyInvocation.MyCommand): "
-                $($LocalizedData.FindNetAdapterMessage)
+                $($script:localizedData.FindNetAdapterMessage)
             ) -join '')
 
         $adapter = Find-NetworkAdapter `
@@ -394,7 +387,7 @@ function Test-TargetResource
 
         # An adapter was found that needs to be changed to the new name
         Write-Verbose -Message ( @("$($MyInvocation.MyCommand): "
-                $($LocalizedData.NetAdapterNameNotMatchMessage -f $adapter.Name, $NewName)
+                $($script:localizedData.NetAdapterNameNotMatchMessage -f $adapter.Name, $NewName)
             ) -join '')
 
         return $false