+++ DSCResources/MSFT_NetAdapterState/MSFT_NetAdapterState.psm1
@@ -5,15 +5,8 @@ Import-Module -Name (Join-Path -Path $modulePath `
         -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
             -ChildPath 'NetworkingDsc.Common.psm1'))
 
-# Import the Networking Resource Helper Module
-Import-Module -Name (Join-Path -Path $modulePath `
-        -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
-            -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
-
 # Import Localization Strings
-$localizedData = Get-LocalizedData `
-    -ResourceName 'MSFT_NetAdapterState' `
-    -ResourcePath (Split-Path -Parent $Script:MyInvocation.MyCommand.Path)
+$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_NetAdapterState'
 
 <#
 .SYNOPSIS
@@ -44,7 +37,7 @@ function Get-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $localizedData.CheckingNetAdapterMessage
+            $script:localizedData.CheckingNetAdapterMessage
         ) -join '')
 
     try
@@ -55,7 +48,7 @@ function Get-TargetResource
     {
         Write-Warning -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $LocalizedData.NetAdapterNotFoundMessage -f $Name
+            $script:localizedData.NetAdapterNotFoundMessage -f $Name
         ) -join '')
     }
 
@@ -63,11 +56,13 @@ function Get-TargetResource
     {
         Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($LocalizedData.NetAdapterTestingStateMessage -f $Name)
+                $($script:localizedData.NetAdapterTestingStateMessage -f $Name)
             ) -join '')
 
-        # Using NET_IF_ADMIN_STATUS as documented here:
-        # https://docs.microsoft.com/en-us/windows/desktop/api/ifdef/ne-ifdef-net_if_admin_status
+        <#
+            Using NET_IF_ADMIN_STATUS as documented here:
+            https://docs.microsoft.com/en-us/windows/desktop/api/ifdef/ne-ifdef-net_if_admin_status
+        #>
 
         $enabled  = [Microsoft.PowerShell.Cmdletization.GeneratedTypes.NetAdapter.NET_IF_ADMIN_STATUS]::Up
         $disabled = [Microsoft.PowerShell.Cmdletization.GeneratedTypes.NetAdapter.NET_IF_ADMIN_STATUS]::Down
@@ -113,7 +108,7 @@ function Set-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $localizedData.CheckingNetAdapterMessage
+            $script:localizedData.CheckingNetAdapterMessage
         ) -join '')
 
     try
@@ -124,7 +119,7 @@ function Set-TargetResource
     {
         Write-Error -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $LocalizedData.NetAdapterNotFoundMessage -f $Name
+            $script:localizedData.NetAdapterNotFoundMessage -f $Name
         ) -join '')
     }
 
@@ -145,7 +140,7 @@ function Set-TargetResource
         {
             Write-Error -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($LocalizedData.NetAdapterSetFailedMessage -f $Name, $State, $_)
+                $($script:localizedData.NetAdapterSetFailedMessage -f $Name, $State, $_)
             ) -join '')
         }
     }
@@ -179,15 +174,16 @@ function Test-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.NetAdapterTestingStateMessage -f $Name)
+            $($script:localizedData.NetAdapterTestingStateMessage -f $Name)
         ) -join '')
 
     $currentState = Get-TargetResource @PSBoundParameters
+
     if ($currentState)
     {
         Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($localizedData.NetAdapterStateMessage -f $Name, $currentState.State)
+                $($script:localizedData.NetAdapterStateMessage -f $Name, $currentState.State)
             ) -join '')
 
         return $currentState.State -eq $State