+++ DSCResources/MSFT_WaitForNetworkTeam/MSFT_WaitForNetworkTeam.psm1
@@ -10,15 +10,8 @@ Import-Module -Name (Join-Path -Path $modulePath `
         -ChildPath (Join-Path -Path 'NetworkingDsc.Common' `
             -ChildPath 'NetworkingDsc.Common.psm1'))
 
-# Import the Storage Resource Helper Module
-Import-Module -Name (Join-Path -Path $modulePath `
-        -ChildPath (Join-Path -Path 'NetworkingDsc.ResourceHelper' `
-            -ChildPath 'NetworkingDsc.ResourceHelper.psm1'))
-
 # Import Localization Strings
-$localizedData = Get-LocalizedData `
-    -ResourceName 'MSFT_WaitForNetworkTeam' `
-    -ResourcePath (Split-Path -Parent $Script:MyInvocation.MyCommand.Path)
+$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_WaitForNetworkTeam'
 
 <#
     .SYNOPSIS
@@ -40,7 +33,7 @@ function Get-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.GettingWaitForNetworkTeamStatusMessage -f $Name)
+            $($script:localizedData.GettingWaitForNetworkTeamStatusMessage -f $Name)
         ) -join '' )
 
 
@@ -88,7 +81,7 @@ function Set-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.SettingWaitForNetworkTeamStatusMessage -f $Name)
+            $($script:localizedData.SettingWaitForNetworkTeamStatusMessage -f $Name)
         ) -join '' )
 
     $lbfoTeamUp = $false
@@ -101,7 +94,7 @@ function Set-TargetResource
         {
             Write-Verbose -Message ( @(
                     "$($MyInvocation.MyCommand): "
-                    $($localizedData.NetworkTeamUpMessage -f $Name)
+                    $($script:localizedData.NetworkTeamUpMessage -f $Name)
                 ) -join '' )
 
             $lbfoTeamUp = $true
@@ -111,7 +104,7 @@ function Set-TargetResource
         {
             Write-Verbose -Message ( @(
                     "$($MyInvocation.MyCommand): "
-                    $($localizedData.NetworkTeamNotUpRetryingMessage -f $Name, $RetryIntervalSec)
+                    $($script:localizedData.NetworkTeamNotUpRetryingMessage -f $Name, $RetryIntervalSec)
                 ) -join '' )
 
             Start-Sleep -Seconds $RetryIntervalSec
@@ -121,7 +114,7 @@ function Set-TargetResource
     if ($lbfoTeamUp -eq $false)
     {
         New-InvalidOperationException `
-            -Message ($localizedData.NetworkTeamNotUpAfterError -f $Name, $RetryCount)
+            -Message ($script:localizedData.NetworkTeamNotUpAfterError -f $Name, $RetryCount)
     } # if
 } # function Set-TargetResource
 
@@ -159,7 +152,7 @@ function Test-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.TestingWaitForNetworkTeamStatusMessage -f $Name)
+            $($script:localizedData.TestingWaitForNetworkTeamStatusMessage -f $Name)
         ) -join '' )
 
     $lbfoTeamStatus = Get-NetLbfoTeamStatus -Name $Name
@@ -168,7 +161,7 @@ function Test-TargetResource
     {
         Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($localizedData.NetworkTeamUpMessage -f $Name)
+                $($script:localizedData.NetworkTeamUpMessage -f $Name)
             ) -join '' )
 
         return $true
@@ -176,7 +169,7 @@ function Test-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.NetworkTeamNotUpMessage -f $Name)
+            $($script:localizedData.NetworkTeamNotUpMessage -f $Name)
         ) -join '' )
 
     return $false
@@ -212,14 +205,14 @@ function Get-NetLbfoTeamStatus
 
         Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.NetworkTeamFoundMessage -f $Name)
+            $($script:localizedData.NetworkTeamFoundMessage -f $Name)
         ) -join '' )
 
     }
     catch [Microsoft.PowerShell.Cmdletization.Cim.CimJobException]
     {
         New-InvalidOperationException `
-            -Message ($LocalizedData.NetworkTeamNotFoundMessage -f $Name)
+            -Message ($script:localizedData.NetworkTeamNotFoundMessage -f $Name)
     }
 
     return $lbfoTeam.Status